rename virtual id util
