+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/sortPartitions/SortPartitionsOpExecConfig.scala
@@ -6,7 +6,7 @@ import edu.uci.ics.amber.engine.architecture.deploysemantics.deploystrategy.{Rou
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.WorkerLayer
 import edu.uci.ics.amber.engine.common.Constants
 import edu.uci.ics.amber.engine.common.tuple.ITuple
-import edu.uci.ics.amber.engine.common.virtualidentity.util.makeLayer
+import edu.uci.ics.amber.engine.common.virtualidentity.VirtualIdentityUtil.makeLayer
 import edu.uci.ics.amber.engine.common.virtualidentity.{
   ActorVirtualIdentity,
   LayerIdentity,
