+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/cache/CacheSourceOpExecConfig.scala
@@ -5,7 +5,7 @@ import edu.uci.ics.amber.engine.architecture.deploysemantics.deploymentfilter.Fo
 import edu.uci.ics.amber.engine.architecture.deploysemantics.deploystrategy.RandomDeployment
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.WorkerLayer
 import edu.uci.ics.texera.workflow.common.storage.OpResultStorage
-import edu.uci.ics.amber.engine.common.virtualidentity.util.makeLayer
+import edu.uci.ics.amber.engine.common.virtualidentity.VirtualIdentityUtil.makeLayer
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, OperatorIdentity}
 import edu.uci.ics.amber.engine.operators.OpExecConfig
