+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/visualization/AggregatedVizOpExecConfig.scala
@@ -9,7 +9,7 @@ import edu.uci.ics.amber.engine.architecture.deploysemantics.deploystrategy.Roun
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.WorkerLayer
 import edu.uci.ics.amber.engine.architecture.linksemantics.{HashBasedShuffle, OneToOne}
 import edu.uci.ics.amber.engine.common.{Constants, IOperatorExecutor}
-import edu.uci.ics.amber.engine.common.virtualidentity.util.makeLayer
+import edu.uci.ics.amber.engine.common.virtualidentity.VirtualIdentityUtil.makeLayer
 import edu.uci.ics.amber.engine.common.virtualidentity.{
   ActorVirtualIdentity,
   LayerIdentity,
